feat(week3-task6): add pages folder for favorites and home pages
